feat: #1917 Update tests + remove unused files
