+++ packages/developer-portal/src/components/pages/app-detail-preview/client/__tests__/app-content.test.tsx
@@ -2,11 +2,11 @@ import React from 'react'
 import AppContent from '../app-content'
 import { appDetailDataStub } from '@/sagas/__stubs__/app-detail'
 import { shallow } from 'enzyme'
-import { AppDetailDataNotNull } from '@/reducers/client/app-detail'
+import { AppDetailModel } from '@reapit/foundations-ts-definitions'
 
 describe('ClientAppContent', () => {
   it('ClientAppContent - should match snapshot', () => {
-    const wrapper = shallow(<AppContent appDetailData={appDetailDataStub as AppDetailDataNotNull} />)
+    const wrapper = shallow(<AppContent appDetailData={appDetailDataStub as AppDetailModel} />)
     expect(wrapper).toMatchSnapshot()
   })
 })
