+++ packages/developer-portal/src/components/pages/app-detail-preview/client/aside.tsx
@@ -5,16 +5,14 @@ import { ContactDeveloperSection } from './contact-developer-modal'
 import useReactResponsive from '@/components/hooks/use-react-responsive'
 import { History } from 'history'
 import routes from '@/constants/routes'
-// Commenting out for now until we have the API
 import {
-  DeveloperSection /*, DeveloperAboutSection */,
+  DeveloperSection,
   CategorySection,
   DesktopIntegrationSection,
   DirectApiSection,
   BackToAppsSection,
 } from '../common/ui-sections'
 import { useHistory } from 'react-router'
-import WebComponentConfig from './web-component-config-modal'
 
 interface AsideProps {
   appDetailData: AppDetailModel
@@ -28,21 +26,16 @@ export const onBackToAppsButtonClick = (history: History) => {
 }
 
 export const Aside: React.FC<AsideProps> = ({ desktopIntegrationTypes, appDetailData }) => {
-  const { category, developer, telephone, supportEmail, homePage, isDirectApi, isWebComponent } = appDetailData
+  const { category, developer, telephone, supportEmail, homePage, isDirectApi } = appDetailData
   const { isMobile } = useReactResponsive()
   const history = useHistory()
 
   return (
     <FlexContainerBasic flexColumn hasPadding hasBackground isFullHeight={!isMobile}>
       {developer && <DeveloperSection developer={developer} isSidebar />}
-      {/** Placeholder for now until we have the API
-        <DeveloperAboutSection isSidebar />
-       */}
-
       <CategorySection category={category} isSidebar />
       <DesktopIntegrationSection desktopIntegrationTypes={desktopIntegrationTypes} isSidebar />
       <DirectApiSection isDirectApi={isDirectApi} isSidebar />
-      {/* {isWebComponent && <WebComponentConfig />} */}
       <ContactDeveloperSection
         contact={{
           developer,
