--- packages/developer-portal/src/components/pages/app-detail-preview/client/index.ts
@@ -1,1 +1,0 @@
-export { default } from './app-detail'
