--- packages/developer-portal/src/components/pages/app-detail-preview/client/web-component-config-modal/__stubs__/web-component-config.ts
@@ -1,14 +1,0 @@
-import { ReduxState } from '@/types/core'
-
-export const webComponentStub = {
-  client: {
-    webComponent: { loading: false, updating: false, data: null, isShowModal: true, negotiators: {} },
-  },
-  auth: {
-    loginSession: {
-      loginIdentity: {
-        clientId: 'DXX',
-      },
-    },
-  },
-} as ReduxState
