--- packages/developer-portal/src/components/pages/app-detail-preview/client/web-component-config-modal/index.ts
@@ -1,1 +1,0 @@
-export { default } from './client-web-component-config'
